adding artists to concerts, pulling artists with songkick api call

+++ lib/song_kick_wrapper.rb
@@ -8,28 +8,46 @@ class SongKickWrapper
   # end
 
   def get_results(user_name)
-    user_name ||= 'surfario'
+    user = User.find_by_songkick_username(user_name)
+    if user
 
-    Rails.logger.info ">>>>> user: #{user_name ||= 'wtf'}"
-    # username = user.songkick_name
-    remote = Songkickr::Remote.new "#{ENV['SONG_KICK_API_KEY']}"
-    Rails.logger.info ">>>> remote: #{remote.inspect}"
-    results = remote.users_tracked_events_calendar(user_name)
-    Rails.logger.info ">>>>>> results: #{results.inspect}"
-    
-    concerts = []
-    results.results.each do |r|
-      concert = {event: r.event.display_name, venue: r.event.venue.display_name, date: r.event.start}
-      concerts << concert  
-    end
-    
-    if concerts.any?
-      Concert.delete_all
+      # Rails.logger.info ">>>>> user: #{user_name ||= 'wtf'}"
+      remote = Songkickr::Remote.new "#{ENV['SONG_KICK_API_KEY']}"
+      # Rails.logger.info ">>>> remote: #{remote.inspect}"
+      results = remote.users_tracked_events_calendar(user_name)
+      # Rails.logger.info ">>>>>> results: #{results.inspect}"
+      
+      concerts = []
+      results.results.each do |r|
+        concert = {event: r.event.display_name, venue: r.event.venue.display_name, date: r.event.start, artists: r.event.performances.first.artist.display_name}
+        # ** Need to grab all artists! **
+        concerts << concert  
+      end
+      
+      if concerts.any?
+        Concert.delete_all
 
-      concerts.each_with_index do |concert, index|
-        Concert.create(event: concert[:event], venue: concert[:venue], date: concert[:date])
+        concerts.each_with_index do |concert, index|
+          concert = Concert.create(event: concert[:event], venue: concert[:venue], date: concert[:date], artists: concert[:artists])
+        
+
+        # concerts.each_with_index do |concert, index|
+        #   concert = Concert.find_by_event(concert[:event])
+        #   unless concert.present? 
+        #     concert = Concert.create(event: concert[:event], venue: concert[:venue], date: concert[:date])
+        #   end
+          
+          artist = Artist.find_by_name(concert[:artists])
+          unless artist.present?
+            artist = Artist.create(name: concert[:artists])
+          end
+        end  
+
+          # user.concerts <<  concert unless user.concerts.include?(concert)
+        
       end
     end   
+
   end
